requirements.txt
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
contextlib2==0.5.5
Django==2.1.2
djangorestframework==3.9.0
idna==2.7
python-dateutil==2.7.5
pytz==2018.6
requests==2.20.0
six==1.11.0
urllib3==1.24
